+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Ships that collide with the edge of the world will be destroyed.
+
 ### 0.48.0
 
 - Increased bullet lifetime.